fix(Vault): cast balance updates to uint96

+++ test/unit/Vault/Deposit.t.sol
@@ -185,13 +185,13 @@ contract VaultDepositTest is UnitBaseSetup, BrokenToken {
   function testDepositMoreThanMax() external {
     vm.startPrank(alice);
 
-    uint256 _amount = uint256(type(uint112).max) + 1;
+    uint256 _amount = uint256(type(uint96).max) + 1;
 
     underlyingAsset.mint(alice, _amount);
     underlyingAsset.approve(address(vault), type(uint256).max);
 
     vm.expectRevert(
-      abi.encodeWithSelector(Vault.DepositMoreThanMax.selector, alice, _amount, type(uint112).max)
+      abi.encodeWithSelector(Vault.DepositMoreThanMax.selector, alice, _amount, type(uint96).max)
     );
 
     vault.deposit(_amount, alice);
@@ -384,13 +384,13 @@ contract VaultDepositTest is UnitBaseSetup, BrokenToken {
   function testMintMoreThanMax() external {
     vm.startPrank(alice);
 
-    uint256 _amount = uint256(type(uint112).max) + 1;
+    uint256 _amount = uint256(type(uint96).max) + 1;
 
     underlyingAsset.mint(alice, _amount);
     underlyingAsset.approve(address(vault), type(uint256).max);
 
     vm.expectRevert(
-      abi.encodeWithSelector(Vault.MintMoreThanMax.selector, alice, _amount, type(uint112).max)
+      abi.encodeWithSelector(Vault.MintMoreThanMax.selector, alice, _amount, type(uint96).max)
     );
 
     vault.mint(_amount, alice);
